requirements.txt
streamlit
langchain>=0.1.7
faiss-cpu
python-dotenv
pandas
tqdm
numpy
huggingface-hub
transformers
sentence-transformers
google-generativeai
sumy
nltk
scikit-learn
langchain-openai
langchain-text-splitters
langchain-embeddings
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pandas
python-dotenv
huggingface_hub
